% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriveLetters.R
\name{DriveLetters}
\alias{DriveLetters}
\title{Drive Letters}
\usage{
DriveLetters()
}
\value{
Return an array.
}
\description{
\code{DriveLetters} gives an array with the letters of the used computer drives.
}
\examples{
DriveLetters()

}

