% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnMapDrive.R
\name{UnMapDrive}
\alias{UnMapDrive}
\title{Unmap a statistical operation}
\usage{
UnMapDrive(DriveLetter = "Z:")
}
\arguments{
\item{DriveLetter}{is an array with the letter of the drive.}
}
\description{
\code{UnMapDrive} unmaps a drive with its statistical operations.
}
\examples{
UnMapDrive('Z:')

}

