% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriveLetters.R
\name{DriveLetters}
\alias{DriveLetters}
\title{Drive Letters}
\usage{
DriveLetters()
}
\value{
It returns a character vector with components [A-Z]?:.
}
\description{
\code{DriveLetters} returns a character vector with the letters 
of the logical drives in use.
}
\examples{
DriveLetters()

}

