% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapDrive.R
\name{MapDrive}
\alias{MapDrive}
\title{Map a statistical operation}
\usage{
MapDrive(SurveyCode, DriveLetter = "Z:", PassWord)
}
\arguments{
\item{SurveyCode}{is an array with the name of the statistical operation.}

\item{DriveLetter}{is an array with the letter of the drive.}

\item{Password}{is an array. It is different for each statistical operation.}
}
\description{
\code{MapDrive} maps a statistical operation SurveyCode in the drive DriveLetter.
A password is needed.
}
\examples{
MapDrive('E30183','Z:','xxxxxxxx')

}

