% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappingStatus.R
\name{MappingStatus}
\alias{MappingStatus}
\title{Detect if a statistical operation is mapped}
\usage{
MappingStatus(SurveyCode, DriveLetter = "Z:")
}
\arguments{
\item{SurveyCode}{is an array with the name of the statistical operation.}

\item{DriveLetter}{is an array with the letter of the drive.}
}
\value{
A Boolean value (TRUE/FALSE).
}
\description{
\code{MappingStatus} detects if the statistical operation SurveyCode is or not mapped into the Drive DriveLetter.
}
\examples{
MappingStatus('E30183','Z:')

}

