% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoStatus.R
\name{RepoStatus}
\alias{RepoStatus}
\title{Repository Status}
\usage{
RepoStatus(SurveyCode, DriveLetter = "Z:", Units = "Mb", Extended = FALSE,
  n = 6L)
}
\arguments{
\item{SurveyCode}{is an array with the name of the statistical operation.}

\item{DriveLetter}{is an array with the letter of the drive.}

\item{Units}{is an array that tells about the units shown in the output.}

\item{Extended}{is an Boolean that takes TRUE value if a more extended output is desired.}

\item{n}{is an integer that informs about the number of shown rows in the output.}
}
\description{
\code{RepoStatus} gives information about the Status of the mapped repository.
}
\examples{
RepoStatus('E30183', DriveLetter = 'Z:', Units = 'Mb', Extended = FALSE, n = 6L)
 
}

