% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SendReport.R
\name{SendReport}
\alias{SendReport}
\title{Send a report}
\usage{
SendReport(SurveyCode, To, n = 6L, Extended = FALSE, Units = "Mb")
}
\arguments{
\item{SurveyCode}{is an array with the name of the statistical operation.}

\item{To}{is an array with the destination email address.}

\item{n}{is an integer that informs about the number of shown rows in the output.}

\item{Extended}{is an Boolean that takes TRUE value if a more extended output is desired.}

\item{Units}{is an array that tells about the units shown in the output.}
}
\description{
\code{SendReport} sends a report to an email address containing information about the repository state.
}
\examples{
SendReport('E30183', 'email@example.com', n = 6L, Extended = FALSE, Units = 'Mb')
 
}

