% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnMapDrive.R
\name{UnMapDrive}
\alias{UnMapDrive}
\title{Unmap a statistical operation}
\usage{
UnMapDrive(DriveLetter = "Z:", force = FALSE)
}
\arguments{
\item{DriveLetter}{Character vector of length 1 with the letter of the logical drive.}

\item{force}{Logical vector of length 1 specifying whether to force the unmapping or not (default
value: FALSE)}
}
\description{
\code{UnMapDrive} unmaps a logical drive.
}
\examples{
UnMapDrive('Z:')

}
